+++ Source/ISHDeploy/Data/Managers/WebAdministrationManager.cs
@@ -14,7 +14,10 @@
  * limitations under the License.
  */
 ﻿using System;
+﻿using System.IO;
 ﻿using System.Linq;
+﻿using System.Management.Automation;
+﻿using System.Reflection;
 ﻿using ISHDeploy.Data.Exceptions;
 ﻿using Microsoft.Web.Administration;
 using ISHDeploy.Data.Managers.Interfaces;
@@ -97,13 +100,12 @@ namespace ISHDeploy.Data.Managers
         /// <exception cref="WindowsAuthenticationModuleIsNotInstalledException"></exception>
         public void EnableWindowsAuthentication(string webSiteName)
         {
-            // TODO: Add check is feature WindowsAuthentication enable on current operation system for Set-ISHSTSConfiguration  cmdlet.
-            // https://jira.sdl.com/browse/TS-11523
-            //if (IsWindowsAuthenticationFeatureEnabled())
-            //{
+            _logger.WriteVerbose($"Enable WindowsAuthentication for site: `{webSiteName}`");
+            if (IsWindowsAuthenticationFeatureEnabled())
+            {
+                _logger.WriteVerbose("IIS-WindowsAuthentication feature is turned on");
                 using (ServerManager manager = ServerManager.OpenRemote(Environment.MachineName))
                 {
-                    _logger.WriteDebug($"Enable WindowsAuthentication for site: `{webSiteName}`");
 
                     var config = manager.GetApplicationHostConfiguration();
 
@@ -118,14 +120,15 @@ namespace ISHDeploy.Data.Managers
                         "system.webServer/security/authentication/windowsAuthentication",
                         locationPath);
                     windowsAuthenticationSection["enabled"] = true;
+                    _logger.WriteVerbose("WindowsAuthentication has been enabled");
 
                     manager.CommitChanges();
                 }
-            //}
-            //else
-            //{
-            //    throw new WindowsAuthenticationModuleIsNotInstalledException("IIS-WindowsAuthentication feature has not been turned on. You can run command: 'Enable-WindowsOptionalFeature -Online -FeatureName IIS-WindowsAuthentication' to enable it");
-            //}
+            }
+            else
+            {
+                throw new WindowsAuthenticationModuleIsNotInstalledException("IIS-WindowsAuthentication feature has not been turned on. You can run command: 'Enable-WindowsOptionalFeature -Online -FeatureName IIS-WindowsAuthentication' to enable it");
+            }
         }
 
         /// <summary>
@@ -137,7 +140,7 @@ namespace ISHDeploy.Data.Managers
         {
             using (ServerManager manager = ServerManager.OpenRemote(Environment.MachineName))
             {
-                _logger.WriteDebug($"Enable WindowsAuthentication for site: `{webSiteName}`");
+                _logger.WriteVerbose($"Disable WindowsAuthentication for site: `{webSiteName}`");
 
                 var config = manager.GetApplicationHostConfiguration();
                 var locationPath = config.GetLocationPaths().FirstOrDefault(x => x.Contains(webSiteName));
@@ -146,11 +149,7 @@ namespace ISHDeploy.Data.Managers
                     "system.webServer/security/authentication/windowsAuthentication",
                     locationPath);
                 windowsAuthenticationSection["enabled"] = false;
-
-                //var anonymousAuthenticationSection = config.GetSection(
-                //    "system.webServer/security/authentication/anonymousAuthentication",
-                //    locationPath);
-                //anonymousAuthenticationSection["enabled"] = true;
+                _logger.WriteVerbose("WindowsAuthentication has been disabled");
 
                 manager.CommitChanges();
             }
@@ -162,9 +161,38 @@ namespace ISHDeploy.Data.Managers
         /// <returns></returns>
         private bool IsWindowsAuthenticationFeatureEnabled()
         {
-            // TODO: Add check is feature WindowsAuthentication enable on current operation system for Set-ISHSTSConfiguration  cmdlet.
-            // https://jira.sdl.com/browse/TS-11523
-            return true;
+            bool isFeatureEnabled = false;
+
+            _logger.WriteVerbose("Checking whether IIS-WindowsAuthentication feature is turned on or not");
+
+            using (var ps = PowerShell.Create())
+            {
+                // Read Check-WindowsAuthenticationFeatureEnabled.ps1 script
+                using (var resourceReader = Assembly.GetExecutingAssembly().GetManifestResourceStream("ISHDeploy.Data.Resources.Check-WindowsAuthenticationFeatureEnabled.ps1"))
+                {
+                    using (var reader = new StreamReader(resourceReader))
+                    {
+                        ps.AddScript(reader.ReadToEnd());
+                    }
+                }
+
+                foreach (PSObject result in ps.Invoke())
+                {
+                    foreach (var verbose in ps.Streams.Verbose.ReadAll())
+                    {
+                        _logger.WriteVerbose(verbose.Message);
+                    }
+
+                    foreach (var warning in ps.Streams.Warning.ReadAll())
+                    {
+                        _logger.WriteWarning(warning.Message);
+                    }
+
+                    isFeatureEnabled = result != null && bool.Parse(result.ToString());
+                }
+            }
+
+            return isFeatureEnabled;
         }
 
         /// <summary>