+++ Source/ISHDeploy.Documentation/CHANGELOG.md
@@ -13,6 +13,10 @@
     - Security Token Service\Integrating with Security Token Service mentiones how to enable internal user authentication. **[Update]**
 	- Security Token Service\ADFS\Integrating with Security Token Service mentiones how to enable internal user authentication. **[Update]**
 
+1. ISHSTS
+    1. Cmdlets
+	      - Set-ISHSTSConfiguration does check - is Windows Authentication feature installed on environment or not.
+
 ## beta-0.9 (Pre-release)
 
 1. Working with deployments